backend/requirements.txt
flask
flask-cors
gradio_client
gunicorn
Pillow
opencv-python-headless
numpy